feat: edit post

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -82,8 +82,11 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::query()->findOrFail($id);
+        if ($category->posts()->count()) {
+            return redirect()->route('categories.index')->with('error', 'There are posts in this category');
+        }
         $category->delete();
-
         return redirect()->route('categories.index')->with('success', 'Category deleted successfully');
+        
     }
 }
